+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - A join between triples maps with different logical targets might result in quads ending up in wrong targets.
 
+### Added
+- Option to serialize output as Jelly.
+
 ### Changed
 - Updated new RML test cases for IO and Core.
+
+### Removed
+- `QuadStore`: removed deprecated `write` methods.
 
 ## [7.3.3] - 2025-04-07